requirements.txt
flet>=0.25.2
python-dotenv>=1.0.0
requests>=2.28.0
asyncio>=3.4.3
aiogram>=2.13.1